# [PERF] project: split `_compute_task_count`