--- a/addons/project/views/project_project_views.xml
+++ b/addons/project/views/project_project_views.xml
@@ -388,7 +388,6 @@
                     <field name="display_name"/>
                     <field name="partner_id"/>
                     <field name="color"/>
-                    <field name="task_count"/>
                     <field name="open_task_count"/>
                     <field name="milestone_count_reached"/>
                     <field name="milestone_count"/>