## Description
On large database with many tasks, opening the default kanban view of a Projects can quite slow. It's often a pain point for large DBs that relies heavily on Projects for their day-to-day operations.

## Analysis
The majority of the bottleneck comes from the computation of the tasks' count. The views loads both `task_count` and `open_task_count`, which are compute non-stored fields, yet only the `open_task_count` is shown and actually needed. They both share the same compute `_compute_task_count`, which will compute the counts of tasks for the projects visible in the view, including tasks that are in a closed state. This is unnecessary, as the view now actually needs only the open tasks, whch is usually a significantly lower number of tasks for large long living databases.

## Solution
1) Decouple the compute for `task_count` and `open_task_count` into their own computes, this achieves two goals:
  - Computing the count for the open tasks is significantly faster, as we are filtering on the open tasks, which is a highly selective criteria (can also be supported by a good partial index of the form `(project_id) WHERE state = <list_of_open_states>`)
  - Computing the regular `task_count` doesn't filter based on state anymore, making it more straight forward.
2) Remove `task_count` from the view, as it is not necessary at all, and if not present, it's compute won't be recomputed.

The only downside of this approach is when a view necessitate both `open_task_count` **and** `task_count`, in that case both compute will be triggered individually, which is a bit slower than doing a combined `_read_group` and filtering on the `state` as it was done before.

## Benchmark
On a DB with over 1.8M `project.task`, with a default favourite filter (5 projects, 1 of which has 1M tasks, but only 44k are open) on hot loading

|         | Before | After | Speed up |
|---------|--------|-------|----------|
| Timings | 1.4s   | 575ms | 2.4x     |

Same benchmark, but without any favourite filter, hot load

|         | Before | After | Speed up |
|---------|--------|-------|----------|
| Timings | 2.3s   | 1.3s  | 1.75x    |

## Reference
task-3924914

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr