@robodoo 2024-06-28
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/171284.png)](https://mergebot.odoo.com/odoo/odoo/pull/171284)